Move utils functions inside utils.py
